# backend/docs/swagger.json
{
  "basePath": "/api/v1",
  "definitions": {
...
    "controllers.ReversePaymentRequest": {
      "properties": {
        "reason": {
          "type": "string"
        }
      },
      "required": [
        "reason"
      ],
      "type": "object"
    },
    "models.APIResponse": {
      "properties": {
        "data": {},
        "message": {
          "example": "Operation successful",
          "type": "string"
        },
        "status": {
          "example": "success",
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "models.SSOTAccountBalance": {
      "properties": {
        "account_category": {
          "type": "string"
        },
        "account_code": {
          "type": "string"
        },
        "account_id": {
          "type": "integer"
        },
        "account_name": {
          "type": "string"
        },
        "account_type": {
          "type": "string"
        },
        "current_balance": {
          "type": "number"
        },
        "is_active": {
          "type": "boolean"
        },
        "is_header": {
          "type": "boolean"
        },
        "last_transaction_date": {
          "type": "string"
        },
        "last_updated": {
          "type": "string"
        },
        "normal_balance": {
          "type": "string"
        },
        "total_credits": {
          "type": "number"
        },
        "total_debits": {
          "type": "number"
        },
        "transaction_count": {
          "type": "integer"
        }
      },
      "type": "object"
    },
...
    "services.PaginatedJournalResponse": {
      "properties": {
        "data": {
          "items": {
            "$ref": "#/definitions/services.JournalResponse"
          },
          "type": "array"
        },
        "limit": {
          "type": "integer"
        },
        "page": {
          "type": "integer"
        },
        "total": {
          "type": "integer"
        },
        "total_pages": {
          "type": "integer"
        }
      },
      "type": "object"
    },
...
  },
  "host": "localhost:8080",
  "info": {
    "contact": {
      "email": "support@swagger.io",
      "name": "API Support",
      "url": "http://www.swagger.io/support"
    },
    "description": "API untuk aplikasi sistem akuntansi yang komprehensif dengan fitur lengkap manajemen keuangan, inventory, sales, purchases, dan reporting.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "Sistema Akuntansi API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/payments": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "deprecated": true,
        "description": "DEPRECATED: This endpoint may cause double posting. Use SSOT Payment routes instead. Get paginated list of payments with filters",
        "parameters": [
          {
            "description": "Page number",
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "limit",
            "type": "integer"
          },
          {
            "description": "Filter by contact ID",
            "in": "query",
            "name": "contact_id",
            "type": "integer"
          },
          {
            "description": "Filter by status",
            "in": "query",
            "name": "status",
            "type": "string"
          },
          {
            "description": "Filter by payment method",
            "in": "query",
            "name": "method",
            "type": "string"
          },
          {
            "description": "Start date (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "type": "string"
          },
          {
            "description": "End date (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.APIResponse"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "[DEPRECATED] Get payments list",
        "tags": [
          "Deprecated-Payments"
        ]
      }
    },
    "/api/payments/account-balances/real-time": {
      "get": {
        "description": "Retrieves current account balances from SSOT materialized view",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/models.SSOTAccountBalance"
              },
              "type": "array"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Get real-time account balances",
        "tags": [
          "Payment Integration"
        ]
      }
    },
    "/api/payments/account-balances/refresh": {
      "post": {
        "description": "Manually triggers a refresh of the account balances materialized view",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Refresh account balances",
        "tags": [
          "Payment Integration"
        ]
      }
    },
...
    "/api/payments/enhanced-with-journal": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Creates a payment and automatically generates corresponding journal entry",
        "parameters": [
          {
            "description": "Payment with Journal Request",
            "in": "body",
            "name": "payment",
            "required": true,
            "schema": {
              "$ref": "#/definitions/services.PaymentWithJournalRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/services.PaymentWithJournalResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Create payment with journal integration",
        "tags": [
          "Payment Integration"
        ]
      }
    },
...
    "/api/payments/journal-entries": {
      "get": {
        "description": "Retrieves all journal entries related to a specific payment",
        "parameters": [
          {
            "description": "Payment ID",
            "in": "query",
            "name": "payment_id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/services.PaginatedJournalResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Get journal entries for payment",
        "tags": [
          "Payment Integration"
        ]
      }
    },
...
    "/api/payments/preview-journal": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Shows what journal entry would be created for a payment without actually creating it",
        "parameters": [
          {
            "description": "Payment Preview Request",
            "in": "body",
            "name": "payment",
            "required": true,
            "schema": {
              "$ref": "#/definitions/services.PaymentWithJournalRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/services.PaymentJournalResult"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Preview journal entry for payment",
        "tags": [
          "Payment Integration"
        ]
      }
    },
...
    "/api/payments/{id}": {
      "delete": {
        "consumes": [
          "application/json"
        ],
        "description": "Delete payment and reverse if needed (Admin Only)",
        "parameters": [
          {
            "description": "Payment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Deletion reason",
            "in": "body",
            "name": "reason",
            "required": true,
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete payment",
        "tags": [
          "Payments"
        ]
      },
      "get": {
        "consumes": [
          "application/json"
        ],
        "deprecated": true,
        "description": "DEPRECATED: This endpoint may cause double posting. Use SSOT Payment routes instead. Get single payment details",
        "parameters": [
          {
            "description": "Payment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.Payment"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "[DEPRECATED] Get payment by ID",
        "tags": [
          "Deprecated-Payments"
        ]
      }
    },
    "/api/payments/{id}/account-updates": {
      "get": {
        "description": "Retrieves the account balance changes caused by a payment transaction",
        "parameters": [
          {
            "description": "Payment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/services.AccountBalanceUpdate"
              },
              "type": "array"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Get account balance updates from payment",
        "tags": [
          "Payment Integration"
        ]
      }
    },
...
    "/api/payments/{id}/reverse": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Reverses a payment transaction and creates a corresponding reversing journal entry",
        "parameters": [
          {
            "description": "Payment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Reversal Request",
            "in": "body",
            "name": "reversal",
            "required": true,
            "schema": {
              "$ref": "#/definitions/controllers.ReversePaymentRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/services.PaymentWithJournalResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Reverse payment and journal entry",
        "tags": [
          "Payment Integration"
        ]
      }
    },
    "/api/payments/{id}/with-journal": {
      "get": {
        "description": "Retrieves a payment along with its associated journal entry information",
        "parameters": [
          {
            "description": "Payment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/services.PaymentWithJournalResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Get payment with journal details",
        "tags": [
          "Payment Integration"
        ]
      }
    },
...
    "/api/purchases/{id}/payments": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "description": "Get all payments for a specific purchase",
        "parameters": [
          {
            "description": "Purchase ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/models.PurchasePayment"
              },
              "type": "array"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get purchase payments",
        "tags": [
          "Purchases"
        ]
      }
    },
...
    "/api/v1/journals": {
      "get": {
        "description": "Get journal entries with filtering, pagination, and sorting",
        "parameters": [
          {
            "description": "Source type filter",
            "in": "query",
            "name": "source_type",
            "type": "string"
          },
          {
            "description": "Status filter",
            "in": "query",
            "name": "status",
            "type": "string"
          },
          {
            "description": "Date from filter (YYYY-MM-DD)",
            "in": "query",
            "name": "date_from",
            "type": "string"
          },
          {
            "description": "Date to filter (YYYY-MM-DD)",
            "in": "query",
            "name": "date_to",
            "type": "string"
          },
          {
            "description": "Reference search",
            "in": "query",
            "name": "reference",
            "type": "string"
          },
          {
            "default": 1,
            "description": "Page number",
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "default": 20,
            "description": "Items per page",
            "in": "query",
            "name": "limit",
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/services.PaginatedJournalResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Get journal entries list",
        "tags": [
          "Journal"
        ]
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Create a new journal entry with validation and optional auto-posting",
        "parameters": [
          {
            "description": "Journal Entry Request",
            "in": "body",
            "name": "journal",
            "required": true,
            "schema": {
              "$ref": "#/definitions/services.JournalEntryRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/services.JournalResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Create a new journal entry",
        "tags": [
          "Journal"
        ]
      }
    },
...
    "/api/v1/journals/account-balances/refresh": {
      "post": {
        "description": "Manually refresh the materialized view for account balances",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            }
          }
        },
        "summary": "Refresh account balances",
        "tags": [
          "Journal"
        ]
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}